Moving FileScanImportExportRepository.
